Refactor Details into component
Issue: habitDetailModal and identityDetailModal has duplicate code chunks
Solution: Pull the details out as Details component
Risk: No risk
Reviewed by: Jason Ko
