+++ frontend/src/components/common/details/details.module.css
@@ -1,0 +1,64 @@
+@value makerBrown from '../../../common/colors.css';
+@value makerTransLightBlack from '../../../common/colors.css';
+
+.title {
+  font-family: "Dancing Script";
+  font-size: 2.4em;
+  color: makerBrown;
+  padding-left: 20px;
+  position: relative;
+  display: inline-block;
+}
+
+.identity {
+  float: right;
+  font-size: 1.2em;
+  margin-top: 15px;
+  margin-right: 30px;
+  font-family: sans-serif;
+  color: makerBrown;
+}
+
+.content {
+  height: 190px;
+  display: flex;
+  justify-content: center;
+}
+
+.description {
+  margin: 10px 0px;
+  padding: 10px 20px;
+  border-radius: 15px;
+  width: 450px;
+  font-family: sans-serif;
+  font-size: 1.3em;
+  border: 1px dashed makerBrown;
+  overflow-y: scroll;
+}
+
+.buttons {
+  height: 40px;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.buttons button {
+  font-size: 1.5em;
+  height: 40px;
+  padding: 0px 10px;
+  margin: 0 10px;
+  background-color: transparent;
+  border-radius: 15px;
+  border: 2px solid makerBrown;
+  color: makerBrown;
+  transition: all 0.2s;
+}
+
+.buttons button:hover {
+  background-color: makerTransLightBlack;
+}
+
+.buttons button:active {
+  outline: none;
+}
