+++ frontend/src/components/habitDetailModal/habitDetailModal.jsx
@@ -1,21 +1,11 @@
-import React, { useContext } from "react";
-import { CloseModalContext } from "../../context/modalContext";
+import React from "react";
+import Details from "../common/details/details";
 import styles from "./habitDetailModal.module.css";
 
 const HabitDetailModal = ({ currentCard, deleteCard }) => {
-  const closeModal = useContext(CloseModalContext);
-  const { name, identity, description } = currentCard;
   return (
     <div className={styles.container}>
-      <h1 className={styles.title}>{name}</h1>
-      <span className={styles.identity}>{identity}</span>
-      <div className={styles.content}>
-        <p className={styles.description}>{description}</p>
-      </div>
-      <div className={styles.buttons}>
-        <button onClick={closeModal}>OK</button>
-        <button onClick={deleteCard}>Delete</button>
-      </div>
+      <Details type="habit" currentCard={currentCard} deleteCard={deleteCard} />
     </div>
   );
 };
