+++ frontend/src/components/identityDetailModal/identityDetailModal.jsx
@@ -1,20 +1,15 @@
-import React, { useContext } from "react";
-import { CloseModalContext } from "../../context/modalContext";
+import React from "react";
+import Details from "../common/details/details";
 import styles from "./identityDetailModal.module.css";
 
 const IdentityDetailModal = ({ currentCard, deleteCard }) => {
-  const closeModal = useContext(CloseModalContext);
-  const { name, description } = currentCard;
   return (
     <div className={styles.container}>
-      <h1 className={styles.title}>{name}</h1>
-      <div className={styles.content}>
-        <p className={styles.description}>{description}</p>
-      </div>
-      <div className={styles.buttons}>
-        <button onClick={closeModal}>OK</button>
-        <button onClick={deleteCard}>Delete</button>
-      </div>
+      <Details
+        type="identity"
+        currentCard={currentCard}
+        deleteCard={deleteCard}
+      />
     </div>
   );
 };
